requirements.txt
streamlit
sentence-transformers
scikit-learn
requests
python-dotenv
textstat
PyPDF2